+++ src/containers/TagFilterPage.js
@@ -1,18 +1,24 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
+import { withRouter } from 'react-router-dom'
 
 import TagFilter from '../components/TagView'
-import { getProjectsByTag } from '../selectors'
+import { getProjectsByTag, getTagCounters } from '../selectors'
 import log from '../helpers/log'
 import * as uiActionCreators from '../actions/uiActions'
+import { getPageNumber } from '../components/common/pagination/helpers'
 
 class TagFilterPage extends Component {
   render() {
     log('Render the <TagFilterPage> container')
     const {
       tagProjects,
+      total,
       tag,
+      pageNumber,
+      url,
+      itemPerPage,
       isLoggedin,
       uiActions,
       ui,
@@ -21,9 +27,12 @@ class TagFilterPage extends Component {
     return (
       <TagFilter
         projects={tagProjects}
+        total={total}
+        itemPerPage={itemPerPage}
+        url={url}
+        pageNumber={pageNumber}
         graphProjects={graphProjects}
         tag={tag}
-        maxStars={tagProjects.length > 0 ? tagProjects[0].stars : 0}
         isLoggedin={isLoggedin}
         ui={ui}
         uiActions={uiActions}
@@ -35,15 +44,26 @@ class TagFilterPage extends Component {
 function mapStateToProps(sortFilter) {
   return function(state, props) {
     const { entities: { tags }, auth: { username }, ui } = state
+    const { location } = props
+    const pageNumber = getPageNumber(location) || 1
+    const itemPerPage = 50
+    const start = itemPerPage * (pageNumber - 1)
+    const url = location.pathname
 
     const tagId = props.match.params.id
+    const total = getTagCounters(state)[tagId]
     const tagProjects = getProjectsByTag({
       criteria: sortFilter,
       tagId,
-      limit: 100
+      start,
+      limit: itemPerPage
     })(state)
     return {
       tagProjects,
+      total,
+      pageNumber,
+      itemPerPage,
+      url,
       // graphProjects,
       tag: tags[tagId],
       isLoggedin: username !== '',
@@ -61,4 +81,6 @@ function mapDispatchToProps(dispatch) {
 }
 
 export default sortFilter =>
-  connect(mapStateToProps(sortFilter), mapDispatchToProps)(TagFilterPage)
+  withRouter(
+    connect(mapStateToProps(sortFilter), mapDispatchToProps)(TagFilterPage)
+  )